Added functions to sync timestamps and updated documentation
Functions have been added so the first measurement will occur on a set interval with respect to the start of an hour e.g 15m measurements occur at minutes 0, 15, 30, 45. This may need to be revised to include offsets but it's there for now. The script now sleeps for the appropriate time between measurements to ensure no drift in seconds over time. Documentation updated.