usb-permissions.sh info
Updated README.md to include more information on the autorun features of GCARE-extras.sh, as well as an extra optional setup step utilising usb-permissions.sh

Modified usb-permissions.sh as the formatting for fstab was incorrect, yet data was still logged to usb stick without permissions issue. Either incorrect format still worked or the problem was solved with another bugfix and permissions was no longer an issue? Either way, code works without su permissions now so it's no longer using a hammer to crack a nut

Also bringing the code in to beta given 6 devices are in active deployment and working without issue